# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusMart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: campus
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular]
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List student housing listings
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: campus
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Rooms list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create room listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/wishlist:
    get:
      operationId: getWishlist
      tags: [users]
      summary: Get user wishlist
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: toggleWishlist
      tags: [users]
      summary: Toggle product in wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistToggleRequest"
      responses:
        "200":
          description: Wishlist updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistToggleResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        phone:
          type: string
        username:
          type: string
        campus:
          type: string
        avatarUrl:
          type: string
        rating:
          type: number
        totalSales:
          type: number
        totalListings:
          type: number
        createdAt:
          type: string
      required:
        - id
        - email
        - username
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
        phone:
          type: string
        campus:
          type: string
        avatarUrl:
          type: string
    Product:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        category:
          type: string
        condition:
          type: string
          enum: [new, like_new, good, fair]
        campus:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: number
        status:
          type: string
          enum: [active, sold, paused]
        badge:
          type: string
          enum: [HOT, NEW, SALE]
        sellerId:
          type: string
        sellerUsername:
          type: string
        sellerAvatar:
          type: string
        sellerRating:
          type: number
        isWishlisted:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - price
        - category
        - campus
    CreateProductRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        category:
          type: string
        condition:
          type: string
        campus:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: number
      required:
        - title
        - price
        - category
        - condition
        - campus
    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - products
        - total
...
    Room:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [bedsitter, single, one_bedroom, two_bedroom, hostel]
        monthlyRent:
          type: number
        campus:
          type: string
        distanceToCampus:
          type: string
        images:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        available:
          type: boolean
        landlordName:
          type: string
        landlordPhone:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - type
        - monthlyRent
        - campus
        - available
    CreateRoomRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        monthlyRent:
          type: number
        campus:
          type: string
        distanceToCampus:
          type: string
        images:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        landlordPhone:
          type: string
      required:
        - title
        - type
        - monthlyRent
        - campus
...
    WishlistToggleRequest:
      type: object
      properties:
        productId:
          type: string
      required:
        - productId
    WishlistToggleResponse:
      type: object
      properties:
        wishlisted:
          type: boolean
        productId:
          type: string
      required:
        - wishlisted
        - productId