# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusMart API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  title: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category: string;
  condition: string;
  campus: string;
  images?: string[];
  stock?: number;
}