# lib/api-zod/src/generated/types/createRoomRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusMart API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRoomRequest {
  title: string;
  description?: string;
  type: string;
  monthlyRent: number;
  campus: string;
  distanceToCampus?: string;
  images?: string[];
  amenities?: string[];
  landlordPhone?: string;
}