# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusMart API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductBadge } from "./productBadge";
import type { ProductCondition } from "./productCondition";
import type { ProductStatus } from "./productStatus";

export interface Product {
  id: string;
  title: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category: string;
  condition?: ProductCondition;
  campus: string;
  images?: string[];
  stock?: number;
  status?: ProductStatus;
  badge?: ProductBadge;
  sellerId?: string;
  sellerUsername?: string;
  sellerAvatar?: string;
  sellerRating?: number;
  isWishlisted?: boolean;
  createdAt?: string;
}