# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusMart API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoomType } from "./roomType";

export interface Room {
  id: string;
  title: string;
  description?: string;
  type: RoomType;
  monthlyRent: number;
  campus: string;
  distanceToCampus?: string;
  images?: string[];
  amenities?: string[];
  available: boolean;
  landlordName?: string;
  landlordPhone?: string;
  createdAt?: string;
}